# swagger.yaml
swagger: "2.0"
info:
  description: "API para gestionar posts."
  version: "1.0.0"
  title: "API de Posts"

# Host that the API is served from
host: 'localhost:22434'
basePath: '/'
schemes:
- 'http'

paths:
  /posts:
    get:
      summary: "Obtiene una lista de todos los posts."
      operationId: "getAllPosts"
      produces:
      - "application/json"
      responses:
        200:
          description: "Lista de posts"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Post"
        500:
          description: "Error del servidor"

    post:
      summary: "Añade un nuevo post"
      operationId: "addPost"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objeto Post que necesita ser añadido"
        required: true
        schema:
          $ref: "#/definitions/Post"
      responses:
        201:
          description: "Post creado"
        500:
          description: "Error del servidor"

  /posts/{id}:
    get:
      summary: "Obtiene un post por su ID"
      operationId: "getPostById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID del post a retornar"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Detalle del post"
          schema:
            $ref: "#/definitions/Post"
        404:
          description: "Post no encontrado"
        500:
          description: "Error del servidor"

    put:
      summary: "Actualiza un post existente"
      operationId: "updatePostById"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Objeto Post con nueva información"
        required: true
        schema:
          $ref: "#/definitions/Post"
      responses:
        200:
          description: "Post actualizado"
        404:
          description: "Post no encontrado"
        500:
          description: "Error del servidor"

    delete:
      summary: "Elimina un post"
      operationId: "deletePostById"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Post eliminado"
        404:
          description: "Post no encontrado"
        500:
          description: "Error del servidor"

definitions:
  Post:
    type: "object"
    required:
    - "nombre"
    - "posicion"
    properties:
      id:
        type: "integer"
        format: "int64"
      nombre:
        type: "string"
      posicion:
        type: "string"
      posicion2:
        type: "string"
      numero:
        type: "integer"
      equipo:
        type: "string"
      descripcion:
        type: "string"
      imagen:
        type: "string"
      supertecnica:
        type: "string"